+++ noodles-gff/CHANGELOG.md
@@ -9,7 +9,7 @@
 
   * gff/io/writer/line/record: Percent-encode source and type ([#407]).
 
-  * gff/record: Percent-decode reference sequence name ([#407]).
+  * gff/record: Percent-decode reference sequence name and source ([#407]).
 
 [#407]: https://github.com/zaeleus/noodles/pull/407